requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
tensorrt>=8.6.0
opencv-python>=4.8.0
numpy>=1.24.0
pycuda>=2023.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
kaggle>=1.5.12
albumentations>=1.3.1
cityscapesscripts>=2.2.0
tensorboard>=2.15.0
pycocotools>=2.0.6
onnxruntime>=1.14.0
pycuda>=2022.1
